requirements.txt
aws-cdk.core==1.91.0

aws-cdk.aws-lambda==1.91.0
aws_cdk.aws_lambda_python==1.91.0
aws_cdk.aws_dynamodb==1.91.0

aws_lambda_powertools

# テスト
pytest
boto3
moto
